Backend/requirements.txt
fastapi
uvicorn
openai
pinecone-client
langchain
pydantic
python-dotenv